Update CheckOutHandler.cs
Added methods:
-RevenueQuarterYearIfStatement
- DetermineQuarter

Edited method:
- CheckOut

RevenueQuarterYearIfStatement:
This method will check if the movie exists inside the RevenuePerShow and if it does it will only add the money that the user now is spending.
Else if it doesnt it will bring all the details like title what quarter of the year the revenue is made and the year etc. so you can easily see what movie made how much and when.

DetermineQuarter:
This will determine the month / quarter it was shown as a show and place it in the json file.

CheckOut:
Now checks if its only number or '-' '/' where its needed.
If its a valid month I will add tonight. Not currently.